add more unit tests for spec validation
