+++ api/internal/endpoints/spec_validation_resource.go
