+++ api/internal/models/testdata/sample-invalid-json.json
@@ -1,0 +1,1 @@
+sample invalid json
