+++ api/internal/models/testdata/sample-invalid-yaml.yaml
@@ -1,0 +1,1 @@
+sample invalid yaml
